검색 방법 및 중복 push 방지
